Animal Races: Clan of the Raptor
Have a third party product here from Eric Morton Presents that I've created and wanted to share with the community, so I've got this and two other files this one references that I was hoping to share. This one contains information specifically related to the Clan of the Raptor product. If there are any improvements you would suggest, feel free to let me know.

I will note that I am unsure how clean or efficient the Owl Clan Heritage racial trait scripts are, but I included a note with it, explaining how the author intended these abilities to work, and the way I was best able to implement it.